backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.28.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0